Moved hdl/common/include/base into base/include/base
